requirements.txt
# Lightning Fast Face Matcher API
# Ultra-fast, minimal face matching with 7000 faces pre-loaded
#
# Copyright (c) 2025 Arnav Angarkar. All rights reserved.
# Author: Arnav Angarkar
#
# PROPRIETARY AND CONFIDENTIAL
# This software and its documentation are proprietary to Arnav Angarkar.
# No part of this software may be copied, reproduced, distributed, transmitted,
# transcribed, stored in a retrieval system, or translated into any human or
# computer language, in any form or by any means, electronic, mechanical,
# magnetic, optical, chemical, manual, or otherwise, without the express
# written permission of Arnav Angarkar.
#
# Unauthorized copying, distribution, or use of this software is strictly
# prohibited and may result in severe civil and criminal penalties.

fastapi==0.115.12
uvicorn==0.34.3
deepface==0.0.93
faiss-cpu==1.9.0.post1
numpy==1.26.4
pillow==11.1.0
tensorflow==2.19.0
tf_keras==2.19.0
python-multipart==0.0.20
requests==2.32.3
